Delete console related stuff, reorg makefile to generate libdriver.a
