+++ src/mainlib.h
@@ -1,0 +1,3 @@
+// Init everything driver needs.
+struct event_base* init_main(int argc, char **argv);
+void setup_signal_handlers();
